requirements.txt
streamlit
playwright
playwright-stealth
google-generativeai
moviepy
PyPDF2
requests
python-dotenv
